refactor: rename New suffix to V2
