+++ crates/hyperswitch_domain_models/src/lib.rs
@@ -6,7 +6,7 @@ pub mod payments;
 #[cfg(feature = "payouts")]
 pub mod payouts;
 pub mod router_data;
-pub mod router_data_new;
+pub mod router_data_v2;
 pub mod router_flow_types;
 pub mod router_request_types;
 pub mod router_response_types;
