+++ crates/router/src/core/payments.rs
@@ -1,6 +1,6 @@
 pub mod access_token;
 pub mod conditional_configs;
-pub mod connector_integration_new_impls;
+pub mod connector_integration_v2_impls;
 pub mod customers;
 pub mod flows;
 pub mod helpers;
