+++ crates/router/src/types/api/files.rs
@@ -3,7 +3,7 @@ pub use hyperswitch_domain_models::router_flow_types::files::{Retrieve, Upload};
 use masking::{Deserialize, Serialize};
 use serde_with::serde_as;
 
-pub use super::files_new::*;
+pub use super::files_v2::*;
 use super::ConnectorCommon;
 use crate::{
     core::errors,
