+++ crates/router/src/types/api/payouts.rs
@@ -8,7 +8,7 @@ pub use hyperswitch_domain_models::router_flow_types::payouts::{
     PoCancel, PoCreate, PoEligibility, PoFulfill, PoQuote, PoRecipient, PoRecipientAccount,
 };
 
-pub use super::payouts_new::*;
+pub use super::payouts_v2::*;
 use crate::{services::api, types};
 
 pub trait PayoutCancel:
